requirements.txt
# general  modules
django >3.2,<3.3
Pillow
djangorestframework
drf-nested-routers

# env controlls
python-decouple

# third party modules
markdown       
django-filter  
drf-yasg[validation]
djangorestframework-simplejwt
django-cors-headers
django-robots
django-debug-toolbar

# deployment module
gunicorn

# refactoring and tests
black
flake8
pytest
pytest-django
Faker

# background proccess and cache
celery
django-redis

# database client
psycopg2-binary

